+++ src/main.cpp
@@ -3,13 +3,11 @@
 #include <sstream>
 #include <string>
 #include <array>
-#include <variant>
 #include <vector>
-#include <optional>
-#include <unordered_map>
 
 #include <editline/readline.h>
 
+#include "main.h"
 
 const int EXIT_ERROR = 65;
 const int EXIT_OK = 0;
@@ -25,9 +23,6 @@ void error(int line, std::string message){
   report(line, std::string{""}, message);
 }
 
-
-typedef std::variant<bool, int64_t, long double, std::string> NonNullLiteral;
-typedef std::optional<NonNullLiteral> Literal;
 
 std::string repr(Literal l){
   /**
@@ -49,86 +44,9 @@ std::string repr(Literal l){
 
 int64_t ZERO = 0;
 
-enum TokenType {
-  // Single-character tokens.
-  LEFT_PAREN, RIGHT_PAREN, LEFT_BRACE, RIGHT_BRACE,
-  COMMA, DOT, MINUS, PLUS, SEMICOLON, SLASH, STAR,
-
-  // One or two character tokens.
-  BANG, BANG_EQUAL,
-  EQUAL, EQUAL_EQUAL,
-  GREATER, GREATER_EQUAL,
-  LESS, LESS_EQUAL,
-
-  // Literals.
-  IDENTIFIER, STRING, NUMBER,
-
-  // Keywords.
-  AND, KW_CLASS, ELSE, FALSE, FUN, KW_FOR, KW_IF, NIL, OR,
-  PRINT, RETURN, SUPER, THIS, TRUE, VAR, WHILE,
-
-  KW_EOF
-};
-// https://stackoverflow.com/a/6281535/1048464
-static std::array<std::string, 39> tokenTypeStrings {
-  // Single-character tokens.
-  "LEFT_PAREN", "RIGHT_PAREN", "LEFT_BRACE", "RIGHT_BRACE",
-  "COMMA", "DOT", "MINUS", "PLUS", "SEMICOLON", "SLASH", "STAR",
-
-  // One or two character tokens.
-  "BANG", "BANG_EQUAL",
-  "EQUAL", "EQUAL_EQUAL",
-  "GREATER", "GREATER_EQUAL",
-  "LESS", "LESS_EQUAL",
-
-  // Literals.
-  "IDENTIFIER", "STRING", "NUMBER",
-
-  // Keywords.
-  "AND", "KW_CLASS", "ELSE", "FALSE", "FUN", "KW_FOR", "KW_IF", "NIL", "OR",
-  "PRINT", "RETURN", "SUPER", "THIS", "TRUE", "VAR", "WHILE",
-
-  "KW_EOF"
-};
-struct Token {
-  TokenType type;
-  std::string lexeme;
-  Literal literal;
-  size_t line;
-  std::string toString();
-  // TODO: Constructor
-};
-
 std::string Token::toString(){
   return std::string {tokenTypeStrings[type] + ": " + lexeme + " " + repr(literal)};
 }
-
-struct Scanner {
-  std::vector<Token> tokens;
-  Scanner(std::string);
-  std::vector<Token> scanTokens();
-
-  private:
-  std::string source;
-  size_t start;
-  size_t current;
-  size_t line;
-
-  bool isAtEnd();
-  char advance();
-  void scanToken();
-  void addToken(TokenType);
-  void addToken(TokenType, Literal);
-  bool match(char);
-  char peek();
-  char peekNext();
-  void scanString();
-  bool isDigit(char);
-  bool isAlpha(char);
-  bool isAlphaNumeric(char);
-  void number();
-  void identifier();
-};
 
 Scanner::Scanner(std::string src){
   start = 0, current = 0, line = 0;
@@ -276,25 +194,6 @@ void Scanner::scanString(){
   // Exclude quotation marks from the string literal
   addToken(STRING, source.substr(start + 2, current - (start + 1)));
 }
-
-static std::unordered_map<std::string, TokenType> keywords {
-  {"and",    AND},
-  {"class",  KW_CLASS},
-  {"else",   ELSE},
-  {"false",  FALSE},
-  {"for",    KW_FOR},
-  {"fun",    FUN},
-  {"if",     KW_IF},
-  {"nil",    NIL},
-  {"or",     OR},
-  {"print",  PRINT},
-  {"return", RETURN},
-  {"super",  SUPER},
-  {"this",   THIS},
-  {"true",   TRUE},
-  {"var",    VAR},
-  {"while",  WHILE},
-};
 
 
 void Scanner::identifier(){
