+++ src/test.cpp
@@ -1,6 +1,7 @@
 # include "main.cpp"
+# include "main.h"
 
-#include "gtest/gtest.h"
+# include "gtest/gtest.h"
 
 int main(int argc, char **argv) {
   ::testing::InitGoogleTest(&argc, argv);
